[Feature] [Security] implement a flashloan architecture (#12)
This pr introduces a flashloan to the vault. A flashloan allows a beneficial owner of a vault to
"borrow" an asset, perform some action with it, and return it to the vault all in one block.

It is essential to the financial properties of the protocol that the beneficial owner is unable 
to "steal" their asset while an entitlement is in place, and the flashloan is one of the 
high-risk vector for stealing an asset. In particular, many NFTs utilize an approval scheme whereby
the owners of an asset may delegate the ability to transfer it to another address, who can later
(in a different block) transfer the asset on their behalf. If it were possible for an approval to
be granted, the owner would be able to effectively steal the asset from the vault in a later block;
however, in our case we believe this is very unlikely becauase while the owner has control of the asset
it will appear to be "owned" by the flashloanreceiver address, not the actual vault. 

We include a variety of tests with many different flashloan reciever sample contracts here.