[Feature] [Gas] create a multi-asset ERC-721 vault (#22)
This vault takes a similar form to our current ERC-721 vault; however, instead of 
requiring that a unique vault be deployed for each underlying ERC-721 token, this single
vault can support multiple tokenIds each with a distinct entitlement and a beneficial owner.

We believe this will result in substantial gas savings by removing the requirement that a
new vault is deployed for each minted option. Additionally, more care is taken in this
implementation to remove extraneous SSTOREs.

One area where we could get further SSTORE improvements would be by thining the 
entitlement struct that we persist to not include the notion of the vault address, as the
vault adderss is the same for each asset within the vault.